+++ flake.nix
@@ -2,6 +2,7 @@
   inputs.nixpkgs.url = "github:NixOS/nixpkgs";
   inputs.flake-utils.url = "github:numtide/flake-utils";
   inputs.pre-commit-hooks.url = "github:cachix/pre-commit-hooks.nix";
+  inputs.pre-commit-hooks.inputs.nixpkgs.follows = "nixpkgs";
 
   outputs = { self, nixpkgs, flake-utils, pre-commit-hooks }:
     flake-utils.lib.eachDefaultSystem (system:
@@ -15,6 +16,17 @@
             nixfmt.enable = true;
             ormolu.enable = true;
             hpack.enable = true;
+          };
+          tools = {
+            ## This setting specifies which tools to use in the `pre-commit`
+            ## hooks. Since we take our tools (`nixfmt`, `ormolu`, `hpack`) from
+            ## `nixpkgs`, then we can simply make sure that
+            ## `pre-commit-hooks.nix`'s `nixpkgs` input follows ours, so there
+            ## is nothing to see here.
+            ##
+            ## NOTE: Configuring `hpack` here would have no effect. See
+            ## https://github.com/cachix/pre-commit-hooks.nix/issues/255
+            ## for more information.
           };
         };
       in {
@@ -67,10 +79,13 @@
       });
 
   nixConfig = {
-    extra-trusted-substituters =
-      [ "https://tweag-cooked-validators.cachix.org/" ];
+    extra-trusted-substituters = [
+      "https://tweag-cooked-validators.cachix.org/"
+      "https://pre-commit-hooks.cachix.org/"
+    ];
     extra-trusted-public-keys = [
       "tweag-cooked-validators.cachix.org-1:g1TP7YtXjkBGXP/VbSTGBOGONSzdfzYwNJM27bn8pik="
+      "pre-commit-hooks.cachix.org-1:Pkk3Panw5AW24TOv6kz3PvLhlH8puAsJTBbOPmBo7Rc="
     ];
   };
 }