+++ src/Cooked/MockChain/Staged.hs
@@ -37,8 +37,10 @@ import Cooked.MockChain.UtxoState
 import Cooked.Skeleton
 import Cooked.Tweak.Common
 import Data.Default
+import qualified Data.Map as Map
 import qualified Ledger.Slot as Ledger
 import qualified Ledger.Tx as Ledger
+import qualified Ledger.Tx.CardanoAPI as Ledger
 import qualified Ledger.Typed.Scripts as Pl
 import qualified Plutus.V2.Ledger.Api as Pl
 
@@ -144,7 +146,7 @@ instance InterpLtl (UntypedTweak InterpMockChain) MockChainBuiltin InterpMockCha
         StateT [Ltl (UntypedTweak InterpMockChain)] InterpMockChain Ledger.CardanoTx
       interpretAndTell (UntypedTweak now) later = do
         mcst <- lift get
-        let managedTxOuts = utxoIndexToTxOutMap . mcstIndex $ mcst
+        let managedTxOuts = getIndex . mcstIndex $ mcst
             managedDatums = mcstDatums mcst
         (_, skel') <- lift $ runTweakInChain now skel
         lift $
@@ -152,14 +154,14 @@ instance InterpLtl (UntypedTweak InterpMockChain) MockChainBuiltin InterpMockCha
             tell $
               MockChainLog
                 [ MCLogSubmittedTxSkel
-                    (SkelContext managedTxOuts managedDatums)
+                    (SkelContext (txOutV2FromLedger <$> managedTxOuts) $ Map.map fst managedDatums)
                     skel'
                 ]
         tx <- validateTxSkel skel'
         lift $
           lift $
             tell $
-              MockChainLog [MCLogNewTx (Ledger.getCardanoTxId tx)]
+              MockChainLog [MCLogNewTx (Ledger.fromCardanoTxId $ Ledger.getCardanoTxId tx)]
         put later
         return tx
   interpBuiltin (TxOutByRefLedger o) = txOutByRefLedger o