+++ src/Cooked/Pretty/Cooked.hs
@@ -176,13 +176,13 @@ instance PrettyCooked MockChainLog where
             : entries
           )
           | pcOptPrintTxHashes opts =
-            go
-              ( "Validated"
-                  <+> PP.parens ("TxId:" <+> prettyCookedOpt opts txId)
-                  <+> prettyTxSkel opts skelContext skel :
-                acc
-              )
-              entries
+              go
+                ( "Validated"
+                    <+> PP.parens ("TxId:" <+> prettyCookedOpt opts txId)
+                    <+> prettyTxSkel opts skelContext skel
+                    : acc
+                )
+                entries
           | otherwise = go ("Validated" <+> prettyTxSkel opts skelContext skel : acc) entries
       go
         acc
@@ -261,22 +261,22 @@ prettyTxSkelOut opts (Pays output) =
   prettyItemize
     ("Pays to" <+> prettyCookedOpt opts (outputAddress output))
     "-"
-    ( prettyCookedOpt opts (outputValue output) :
-      catMaybes
-        [ case outputOutputDatum output of
-            Pl.OutputDatum _datum ->
-              Just $
-                "Datum (inlined):"
-                  <+> (PP.align . prettyCookedOpt opts)
-                    (output ^. outputDatumL)
-            Pl.OutputDatumHash _datum ->
-              Just $
-                "Datum (hashed):"
-                  <+> (PP.align . prettyCookedOpt opts)
-                    (output ^. outputDatumL)
-            Pl.NoOutputDatum -> Nothing,
-          getReferenceScriptDoc opts output
-        ]
+    ( prettyCookedOpt opts (outputValue output)
+        : catMaybes
+          [ case outputOutputDatum output of
+              Pl.OutputDatum _datum ->
+                Just $
+                  "Datum (inlined):"
+                    <+> (PP.align . prettyCookedOpt opts)
+                      (output ^. outputDatumL)
+              Pl.OutputDatumHash _datum ->
+                Just $
+                  "Datum (hashed):"
+                    <+> (PP.align . prettyCookedOpt opts)
+                      (output ^. outputDatumL)
+              Pl.NoOutputDatum -> Nothing,
+            getReferenceScriptDoc opts output
+          ]
     )
 
 prettyTxSkelOutDatumMaybe :: PrettyCookedOpts -> TxSkelOutDatum -> Maybe DocCooked
@@ -309,12 +309,12 @@ prettyTxSkelIn opts skelContext (txOutRef, txSkelRedeemer) = do
     prettyItemize
       ("Spends from" <+> ownerDoc)
       "-"
-      ( prettyCookedOpt opts (outputValue output) :
-        catMaybes
-          [ redeemerDoc,
-            prettyTxSkelOutDatumMaybe opts txSkelOutDatum,
-            getReferenceScriptDoc opts output
-          ]
+      ( prettyCookedOpt opts (outputValue output)
+          : catMaybes
+            [ redeemerDoc,
+              prettyTxSkelOutDatumMaybe opts txSkelOutDatum,
+              getReferenceScriptDoc opts output
+            ]
       )
 
 prettyTxSkelInReference :: PrettyCookedOpts -> SkelContext -> Pl.TxOutRef -> Maybe DocCooked
@@ -324,11 +324,11 @@ prettyTxSkelInReference opts skelContext txOutRef = do
     prettyItemize
       ("References output from" <+> prettyCookedOpt opts (outputAddress output))
       "-"
-      ( prettyCookedOpt opts (outputValue output) :
-        catMaybes
-          [ prettyTxSkelOutDatumMaybe opts txSkelOutDatum,
-            getReferenceScriptDoc opts output
-          ]
+      ( prettyCookedOpt opts (outputValue output)
+          : catMaybes
+            [ prettyTxSkelOutDatumMaybe opts txSkelOutDatum,
+              getReferenceScriptDoc opts output
+            ]
       )
 
 getReferenceScriptDoc :: (IsAbstractOutput output, ToScriptHash (ReferenceScriptType output)) => PrettyCookedOpts -> output -> Maybe DocCooked
@@ -364,6 +364,7 @@ mPrettyTxOpts
       txOptBalance,
       txOptBalanceOutputPolicy,
       txOptBalanceWallet,
+      txOptBalancingUtxos,
       txOptEmulatorParamsModification
     } =
     prettyItemizeNonEmpty "Options:" "-" $
@@ -373,6 +374,7 @@ mPrettyTxOpts
           prettyIfNot True prettyBalance txOptBalance,
           prettyIfNot def prettyBalanceOutputPolicy txOptBalanceOutputPolicy,
           prettyIfNot def prettyBalanceWallet txOptBalanceWallet,
+          prettyIfNot def prettyBalancingUtxos txOptBalancingUtxos,
           prettyIfNot [] prettyUnsafeModTx txOptUnsafeModTx,
           prettyIfNot def prettyEmulatorParamsModification txOptEmulatorParamsModification
         ]
@@ -406,6 +408,13 @@ mPrettyTxOpts
       prettyEmulatorParamsModification :: Maybe EmulatorParamsModification -> DocCooked
       prettyEmulatorParamsModification Nothing = "No modifications of protocol paramters"
       prettyEmulatorParamsModification Just {} = "With modifications of protocol parameters"
+      prettyBalancingUtxos :: BalancingUtxos -> DocCooked
+      prettyBalancingUtxos BalancingUtxosAll = "Balance with all UTxOs of the balancing wallet"
+      prettyBalancingUtxos BalancingUtxosDatumless = "Balance with datumless UTxOs of the balancing wallet"
+      prettyBalancingUtxos (BalancingUtxosAllowlist txOutRefs) =
+        prettyItemize "Only balance with UTxOs of the balancing wallet among:" "-" (prettyCookedOpt opts <$> txOutRefs)
+      prettyBalancingUtxos (BalancingUtxosBlocklist txOutRefs) =
+        prettyItemize "Do not balance with UTxOs among:" "-" (prettyCookedOpt opts <$> txOutRefs)
 
 -- * Pretty-printing