+++ tests/Cooked/InlineDatumsSpec.hs
@@ -224,13 +224,13 @@ tests =
                     testSucceeds def $
                       spendOutputTestTrace True (inputDatumValidator True),
                   testCase "...and gets a datum hash, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       spendOutputTestTrace False (inputDatumValidator True)
                 ],
               testGroup
                 "validator expects a datum hash..."
                 [ testCase "...and gets an inline datum, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       spendOutputTestTrace True (inputDatumValidator False),
                   testCase "...and gets a datum hash, expecting success" $
                     testSucceeds def $
@@ -245,31 +245,31 @@ tests =
                     testSucceeds def $
                       continuingOutputTestTrace Datum (outputDatumValidator Datum),
                   testCase "...and gets an inline datum, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace Inline (outputDatumValidator Datum),
                   testCase "...and gets a datum hash, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace OnlyHash (outputDatumValidator Datum)
                 ],
               testGroup
                 "validator expects an inline datum..."
                 [ testCase "...and gets a regular datum, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace Datum (outputDatumValidator Inline),
                   testCase "...and gets an inline datum, expecting success" $
                     testSucceeds def $
                       continuingOutputTestTrace Inline (outputDatumValidator Inline),
                   testCase "...and gets a datum hash, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace OnlyHash (outputDatumValidator Inline)
                 ],
               testGroup
                 "validator expects a datum hash..."
                 [ testCase "...and gets a regular datum, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace Datum (outputDatumValidator OnlyHash),
                   testCase "...and gets an inline datum, expecting script failure" $
-                    testFailsFrom' def (isCekEvaluationFailure def) def $
+                    testFails def (isCekEvaluationFailure def) $
                       continuingOutputTestTrace Inline (outputDatumValidator OnlyHash),
                   testCase "...and gets a datum hash, expecting success" $
                     testSucceeds def $