+++ CHANGELOG.md
@@ -18,6 +18,15 @@
   of `BuiltinString`, but potentially useful for "printf-debugging" of scripts.
 - An option `txOptEmulatorParamsModification` to temporarily change protocol
   parameters for balancing and validation of a transaction
+- A function `combineModsTweak` to construct branching tweaks depending on the
+  different combinations of foci of an optic on `TxSkel`
+- New `PrettyCooked` instances for common Plutus types
+- Tweaks on signers in the non-lens tweak API
+- A function `resolveTypedDatum` to recover typed data on UTxOs in
+  `MonadBlockChainBalancing`.
+- An `UtxoSearch` that starts from a list of `TxOutRef`s
+- A transaction option to choose which UTxOs can be spent for balancing
+- Lenses for the fields of `TxOpts`
 
 ### Removed
 
@@ -29,6 +38,25 @@
 
 - `Cooked.holdingInState` is relpaced by `Cooked.holdsInState` which takes an
   address instead of a wallet as argument.
+- Failure testing is slightly modified so that every test has to check that the
+  right error is thrown
+  * `Cooked.testFailsFrom'` is renamed to `Cooked.testFailsFrom`
+  * `Cooked.testFails` is (the new) `Cooked.testFailsFrom` with the default
+    distribution.
+
+  To update their code, users must
+  1. Adapt invokations of `Cooked.testFails` and `Cooked.testFailsFrom` adding
+     a predicate that must hold on the error returned by running the
+     transaction,
+  2. Rename `Cooked.testFailsFrom'` into `Cooked.testFailsFrom`.
+  3. (Bonus) simplify, knowing that ``Cooked.testFailsFrom o x def ==
+     Cooked.testFails o x``
+
+### Fixes
+
+- Add forgotten export of `permanentValue`
+- In `MockChainT`: don't delete data on transaction inputs if there are still
+  UTxOs with that datum around. (See PR #354)
 
 ## [[2.0.0]](https://github.com/tweag/cooked-validators/releases/tag/v2.0.0) - 2023-02-28